fix: remove snackbar context
