+++ README.md
@@ -24,7 +24,6 @@ We believe this is a better approach than the alternatives:
 **Note!** At the moment there exist some parallel solutions, a legacy one which uses react context and a newer one, which we recommend, using hooks (utilizes [zustand](https://github.com/pmndrs/zustand) under the hood). They are:
 
 - `useSharedPortalArea`
-- `useSnackbar`
 
 The solutions based on react context may be removed in an upcoming major release.
