+++ src/contexts/index.ts
@@ -1,6 +1,5 @@
 export * from './Auth'
 export * from './SharedPortalArea'
-export * from './Snackbar'
 export * from './Strings'
 export * from './Theme'
 export * from './Urql'
