+++ src/example-app/App.tsx
@@ -8,16 +8,18 @@ import { ActivityIndicator, Switch } from 'react-native-paper'
 import Animated, { CurvedTransition } from 'react-native-reanimated'
 import { SafeAreaProvider } from 'react-native-safe-area-context'
 
+import { SnackbarPresentationView } from '../components'
 import NativePortal from '../components/NativePortal'
 import DefaultSnackbarComponent from '../components/SnackbarComponent'
 import { StringsProvider } from '../contexts/Strings'
-import { useIsKeyboardShown, useKeyboardHeight, useWillKeyboardBeShown } from '../hooks'
+import {
+  useAddSnackbar, useIsKeyboardShown, useKeyboardHeight, useWillKeyboardBeShown, useSnackbarSettings,
+} from '../hooks'
 import useAlert from '../hooks/useAlert'
 import useConfirm from '../hooks/useConfirm'
 import {
   SharedPortalAreaProvider, SharedPortalPresentationArea, useSharedPortalAreaInsets, useSharedPortalAreaSize,
 } from '../hooks/useSharedPortalArea'
-import { SnackbarPresentationView, useAddSnackbar, useSnackbarSettings } from '../hooks/useSnackbar'
 import Column from '../primitives/Column'
 import Row from '../primitives/Row'
