+++ src/hooks/useAddSnackbar.ts
@@ -1,15 +1,11 @@
-import { useCallback, useContext } from 'react'
+import { useCallback } from 'react'
 
-import { SnackbarContext } from '../contexts/Snackbar'
+import useSnackbarStore from './useSnackbarStore'
 
-import type { SnackbarConfig } from '../contexts/Snackbar'
+import type { SnackbarConfig } from './useSnackbarStore'
 
-/**
- * **Note!** This hook is to be used in conjunction with the Snackbar context, use
- * `useAddSnackbar` from `src/hooks/useSnackbar.tsx` if you are not using the Snackbar context.
- */
 export function useAddSnackbar<TMap extends Record<string, unknown> = Record<string, unknown>, T extends keyof TMap = keyof TMap>(defaultSnackbarConfig?: Omit<SnackbarConfig<TMap, T>, 'title'>) {
-  const { addSnackbar } = useContext(SnackbarContext)
+  const addSnackbar = useSnackbarStore((state) => state.addSnackbar)
 
   return useCallback(function ShowSnackbar<TMapInner extends Record<string, unknown> = TMap, TInner extends keyof TMapInner = keyof TMapInner>(title: string, snackbarConfig?: Omit<SnackbarConfig<TMapInner, TInner>, 'title'>) {
     addSnackbar<TMapInner, TInner>({ ...defaultSnackbarConfig, ...snackbarConfig, title } as SnackbarConfig<TMapInner, TInner>)
