+++ src/hooks/useRemoveSnackbar.ts
@@ -1,0 +1,7 @@
+import useSnackbarStore from './useSnackbarStore'
+
+export function useRemoveSnackbar() {
+  return useSnackbarStore((state) => state.removeSnackbar)
+}
+
+export default useRemoveSnackbar
