+++ src/hooks/useSnackbarWasPresented.ts
@@ -1,0 +1,7 @@
+import useSnackbarStore from './useSnackbarStore'
+
+export function useSnackbarWasPresented() {
+  return useSnackbarStore((state) => state.snackbarWasPresented)
+}
+
+export default useSnackbarWasPresented
