+++ src/hooks/useSnackbarsToShow.ts
@@ -1,0 +1,7 @@
+import useSnackbarStore from './useSnackbarStore'
+
+export function useSnackbarsToShow() {
+  return useSnackbarStore((state) => state.snackbarsToShow)
+}
+
+export default useSnackbarsToShow
